+++ autolens/imaging/model/analysis.py
@@ -1,0 +1,201 @@
+from autoconf import conf
+import autofit as af
+import autoarray as aa
+import autogalaxy as ag
+
+from autolens.analysis.analysis import AnalysisDataset
+from autolens.imaging.model.result import ResultImaging
+from autolens.imaging.model.visualizer import VisualizerImaging
+from autolens.imaging.fit_imaging import FitImaging
+
+from autolens import exc
+
+
+class AnalysisImaging(AnalysisDataset):
+    @property
+    def imaging(self):
+        return self.dataset
+
+    def log_likelihood_function(self, instance):
+        """
+        Determine the fit of a lens galaxy and source galaxy to the imaging in this lens.
+
+        Parameters
+        ----------
+        instance
+            A model instance with attributes
+
+        Returns
+        -------
+        fit : Fit
+            A fractional value indicating how well this model fit and the model imaging itself
+        """
+
+        self.associate_hyper_images(instance=instance)
+        tracer = self.tracer_for_instance(instance=instance)
+
+        self.settings_lens.check_positions_trace_within_threshold_via_tracer(
+            tracer=tracer, positions=self.positions
+        )
+
+        hyper_image_sky = self.hyper_image_sky_for_instance(instance=instance)
+
+        hyper_background_noise = self.hyper_background_noise_for_instance(
+            instance=instance
+        )
+
+        try:
+            return self.fit_imaging_for_tracer(
+                tracer=tracer,
+                hyper_image_sky=hyper_image_sky,
+                hyper_background_noise=hyper_background_noise,
+            ).figure_of_merit
+        except (
+            exc.PixelizationException,
+            exc.InversionException,
+            exc.GridException,
+            OverflowError,
+        ) as e:
+            raise exc.FitException from e
+
+    def fit_imaging_for_tracer(
+        self, tracer, hyper_image_sky, hyper_background_noise, use_hyper_scalings=True
+    ):
+
+        return FitImaging(
+            imaging=self.dataset,
+            tracer=tracer,
+            hyper_image_sky=hyper_image_sky,
+            hyper_background_noise=hyper_background_noise,
+            use_hyper_scaling=use_hyper_scalings,
+            settings_pixelization=self.settings_pixelization,
+            settings_inversion=self.settings_inversion,
+            preloads=self.preloads,
+        )
+
+    def stochastic_log_evidences_for_instance(self, instance):
+
+        instance = self.associate_hyper_images(instance=instance)
+        tracer = self.tracer_for_instance(instance=instance)
+
+        if not tracer.has_pixelization:
+            return
+
+        if not isinstance(
+            tracer.pixelizations_of_planes[-1], aa.pix.VoronoiBrightnessImage
+        ):
+            return
+
+        hyper_image_sky = self.hyper_image_sky_for_instance(instance=instance)
+
+        hyper_background_noise = self.hyper_background_noise_for_instance(
+            instance=instance
+        )
+
+        settings_pixelization = (
+            self.settings_pixelization.settings_with_is_stochastic_true()
+        )
+
+        log_evidences = []
+
+        for i in range(self.settings_lens.stochastic_samples):
+
+            try:
+                log_evidence = FitImaging(
+                    imaging=self.dataset,
+                    tracer=tracer,
+                    hyper_image_sky=hyper_image_sky,
+                    hyper_background_noise=hyper_background_noise,
+                    settings_pixelization=settings_pixelization,
+                    settings_inversion=self.settings_inversion,
+                    preloads=self.preloads,
+                ).log_evidence
+            except (
+                exc.PixelizationException,
+                exc.InversionException,
+                exc.GridException,
+                OverflowError,
+            ) as e:
+                log_evidence = None
+
+            if log_evidence is not None:
+                log_evidences.append(log_evidence)
+
+        return log_evidences
+
+    def visualize(self, paths: af.DirectoryPaths, instance, during_analysis):
+
+        instance = self.associate_hyper_images(instance=instance)
+        tracer = self.tracer_for_instance(instance=instance)
+        hyper_image_sky = self.hyper_image_sky_for_instance(instance=instance)
+        hyper_background_noise = self.hyper_background_noise_for_instance(
+            instance=instance
+        )
+
+        fit = self.fit_imaging_for_tracer(
+            tracer=tracer,
+            hyper_image_sky=hyper_image_sky,
+            hyper_background_noise=hyper_background_noise,
+        )
+
+        visualizer = VisualizerImaging(visualize_path=paths.image_path)
+
+        visualizer.visualize_imaging(imaging=self.imaging)
+        visualizer.visualize_fit_imaging(fit=fit, during_analysis=during_analysis)
+        visualizer.visualize_tracer(
+            tracer=fit.tracer, grid=fit.grid, during_analysis=during_analysis
+        )
+        if fit.inversion is not None:
+            visualizer.visualize_inversion(
+                inversion=fit.inversion, during_analysis=during_analysis
+            )
+
+        visualizer.visualize_hyper_images(
+            hyper_galaxy_image_path_dict=self.hyper_galaxy_image_path_dict,
+            hyper_model_image=self.hyper_model_image,
+            tracer=tracer,
+        )
+
+        if visualizer.plot_fit_no_hyper:
+            fit = self.fit_imaging_for_tracer(
+                tracer=tracer,
+                hyper_image_sky=None,
+                hyper_background_noise=None,
+                use_hyper_scalings=False,
+            )
+
+            visualizer.visualize_fit_imaging(
+                fit=fit, during_analysis=during_analysis, subfolders="fit_no_hyper"
+            )
+
+    def save_results_for_aggregator(
+        self,
+        paths: af.DirectoryPaths,
+        samples: af.OptimizerSamples,
+        model: af.Collection,
+    ):
+
+        pixelization = ag.util.model.pixelization_from(model=model)
+
+        if conf.instance["general"]["hyper"]["stochastic_outputs"]:
+            if ag.util.model.isinstance_or_prior(
+                pixelization, aa.pix.VoronoiBrightnessImage
+            ):
+                self.save_stochastic_outputs(paths=paths, samples=samples)
+
+    def make_result(
+        self, samples: af.PDFSamples, model: af.Collection, search: af.NonLinearSearch
+    ):
+        return ResultImaging(samples=samples, model=model, analysis=self, search=search)
+
+    def save_attributes_for_aggregator(self, paths: af.DirectoryPaths):
+
+        super().save_attributes_for_aggregator(paths=paths)
+
+        paths.save_object("psf", self.dataset.psf_unormalized)
+        paths.save_object("mask", self.dataset.mask)
+        paths.save_object("positions", self.positions)
+        if self.preloads.sparse_grids_of_planes is not None:
+            paths.save_object(
+                "preload_sparse_grids_of_planes", self.preloads.sparse_grids_of_planes
+            )
