+++ test_autolens/imaging/model/test_result_imaging.py
@@ -1,0 +1,8 @@
+import os
+
+directory = os.path.dirname(os.path.realpath(__file__))
+
+
+class TestResultImaging:
+
+    pass
